+++ R/nesting.R
@@ -4,6 +4,10 @@
 #'
 #' @inheritParams e_bar
 #' @param param The nested parameter to add data to.
+#' @param .serie Serie's index to add the data to, if
+#' `NULL` then it is added to all.
+#' @param .data A dataset to use, if none are specified than
+#' the original dataset passed to `e_charts` is used.
 #'
 #' @details For instance, \code{\link{e_funnel}} lets you pass \code{values} and \code{labels}
 #' (from your initial data.frame) which corresponds to \code{name} and \code{value} in the
@@ -53,18 +57,26 @@
 #'     fontStyle
 #'   )
 #' @export
-e_add <- function(e, param, ...) {
+e_add <- function(e, param, ..., .serie = NULL, .data = NULL) {
   if (missing(e) || missing(param)) {
     stop("missing e or param", call. = FALSE)
   }
 
-  for (i in seq_along(e$x$data)) {
-    data <- e$x$data[[i]] |>
+  if(is.null(.data))
+    ds <- e$x$data
+  else
+    ds <- map_grps_(.data, e$x$tl)
+
+  for (i in seq_along(ds)) {
+    data <- ds[[i]] |>
       dplyr::select(...)
 
     data <- apply(data, 1, as.list)
 
     for (j in seq_along(data)) {
+      if(!is.null(.serie) && .serie != j)
+        next
+      
       if (!e$x$tl) {
         e$x$opts$series[[i]]$data[[j]][[param]] <- data[[j]]
       } else {